Backend handles all API calls, frontend reads from the server
Setup the backend to handle all api calls on a seperate database since the backend cant use expo. The front end simply reads from the backend eb and updates the local storage db on an interval. The backend database that queries the api is on a 10 minute interval to minimize api calls and the front end news updating is on a 5 minute interval just to check for changes to update accordingly.