Front end and backend communication success
the front end and backend are communicating between their news tables. I had to set up a different backend_url in news.ts in order for this to work. According to online research. for android emulation, we should be using the backend url: http://10.0.2.2:3001

Once I implemented this and tested it, all error disapeard